develop to main (#45)
* feat: Add detecting mobile or not

user angent를 이용한 함수로 판단하도록 수정

* font 재적용 (#18)


* feat: Add pebble animation in background

* feat: Create GiftBox waiting component

* feat: Add GiftBox compoennt in Participation

* 이벤트 생성 API 호출, 이미지 업로드 중간 구현 (#19)

* api request 명에 맞춰 싱크

* 이벤트 생성 API 연동
- server proxy 추가

* 이미지 업로드 적용
- 로그인 api 변경


* feat: Add open function of GiftBox

* 이미지 업로드 구현 (#21)


* fix: Change height of MobileWrapper

min-height로 잘못 설정된 부분 수정

* fix: Add min-height

* auth header 적용함수 따로 추출

* fix: Add absolute style to block empty height

* feat: Change wrapper size

Wrapper 변경에 따른 사이즈 변경

* feat: Create Particles components

파티클 애니메이션 추가

* 단비님께서 작업해주신 화면까지 새 디자인 적용, 당첨/탈락 이미지 선택 로직 추가 (#27)

* 이벤트 개설 페이지 NEW 디자인 적용

* dialog 임시 적용

* CardListForm 의 아이템 작업
- getImageSource에 path추가하고 이미지 이름만 받기
- Card컴포넌트 bgWhite props 추가

* 이미지 업로드 및 선택 로직 추가


* 메뉴 아이콘 이미지 수정 (#28)

* 이벤트 개설 페이지 NEW 디자인 적용

* dialog 임시 적용

* CardListForm 의 아이템 작업
- getImageSource에 path추가하고 이미지 이름만 받기
- Card컴포넌트 bgWhite props 추가

* 이미지 업로드 및 선택 로직 추가

* 메뉴에 새 아이콘이미지 적용


* 숨겨진 메뉴 height으로 인해 스크롤 생기는 문제 해결

* 폼 관리 로직 리팩터링 (#30)

* FormContext 사용으로 페이지 벗어나도 state 유지

* radioform 리팩터링
- react hook form 구조로 일관되게 통일

* 당첨,탈락 이미지 리스트또한 form으로 관리, 이벤트 개설 api request 수정


* 이벤트 목록 조회 및 이벤트 개설 플로우 완성 (#31)

* 개설한 이벤트 목록 조회 api 연동
- 로그인 화면 height 수정
- api data 타입에 맞춰 response 수정
- 컴포넌트 수정

* 로그인 로직 수정
- data 캐싱 방지
- font preconnect 적용
- api Suspense 적용

* 에러 페이지 선작업

* 에러 페이지 작업

* router 지연로딩 적용

* useMyQuery -> useGetQuery

* 이벤트 개설 완료 페이지 추가 및 이동 로직 연결


* lint 수정

* feat: Change position of particle

* refactor: Remove useless code

* feat: Add remain over 1 day

1일 이상 남은 경우에 대한 시간 표시 수정

* fix: Change second calculate logic

* feat: Add event routing (add event code)

* feat: Use fetch to participateEvent

오류가 있어 임시로 fetch를 쓰도록 변경

* feat: Change API call logic, status check logic

* 이벤트 개설 > 상품 등록 Dialog 로직 추가 (#34)

* Dialog 를 React.Portal로 띄우기

* 상품 등록 로직 구현

* 상품 등록 후 등록Dialog의 입력값 초기화
- mode props 추가
- Dialog에 onOpen props 추가


* feat: Create InvalidCode

event code가 유효하지 않을 때 렌더링할 컴포넌트 생성

* feat: Add useParams hook

* feat: Create Canceled component

취소 상태일때 렌더링할 컴포넌트 추가

* feat: Create checkEventResult

이벤트 결과 조회 API 추가

* feat: Create debounce function

* feat: Add check result handeler in Participation

필요한 데이터들 prop으로 넘기도록 변경

* feat: Add data of event result

* feat: Add layout in InvalidCode

레이아웃 문제 수정

* fix: Change event code param & check status logic

* 링크 복사 기능 구현, 이미지 다운로드 WIP (#35)

* RadioForm에 selectIcon props 추가

* 링크 복사 구현

* (WIP) 이미지 다운로드 기능 구현


* feat: Add Content-Type json

* fix: Change empty check logic

* feat: Add error handle logic & success button

* 이미지 다운로드 구현 (#36)

* post 요청은 onSettle 시에도 초기화

* 상품등록 모달 디자인 업데이트

* 이미지 다운로드 기능 구현

* 이미지 다운로드 수정

* 이벤트 개설 완료 페이지에 복사기능 추가


* 하프레이어 추가 및 당첨/탈락 메시지 선택레이어 구현 (#37)

* dialog 컬러 수정

* halflayer 구현 (WIP)

* 당첨/탈락 선택 레이어 내 RadioForm 구현

* 린트 수정

* overlay 로 인한 zIndex 이슈 수정

* select layer title, list props로 로직 수정

* Empty Event 컴포넌트 추가

* EmptyLayout 추가 및 적용


* 이미지 삭제 구현 (#38)

* 이미지 삭제 구현

* 당첨/탈락 메세지 셀렉터에 선택한 문구 보여주기

* 당첨, 탈락 이미지 삭제 구현


* feat: Add change state when timer is over

* refactor: Remove hardcoding status

* fix: Move waiting over logic to interval

* feat: Create Finished event component

* 이벤트 기간 벗어난 에러 핸들링

* feat: Add event over error handle page

* Form Validation 구현 (#42)

form validation 및 formText 추가

* 이벤트 상세 페이지 구현, Toast 구현 및 적용 (#44)

* 이벤트 상세 페이지 및 조회 API 호출 로직 구현

* 이벤트 상세 페이지 추가

* 이벤트 카드 타이틀 컴포넌트에 ellipsis 적용

* event 상세 조회 api 연동
- react-query refetch 옵션 추가

* toast 구현

* 클라에서 토큰이 만료되어 Cookie에서 값이 없을 시, AuthorizationError throw
- 현재 이벤트 참여 API는 제외되었음 (확인 불가 이슈)

* react-helmet 적용
- title 각 페이지 기능에 맞게 변경
- 이벤트 참여 페이지의 og:title 동적으로 변경 (test)
- Event 참여 API에 react-query 적용 (refetch 확인)

* 로그아웃 후 isLogined와 pathname에 따라 로그인 화면으로 이동

* ROUTE.EVENT.OVER 제거

* 이벤트 페이지 리팩토링 완료
- 로그아웃 오류 수정
- 500 이상 에러는 무조건 ErrorBoundary 로 이동

+++ src/util/client.ts
@@ -58,6 +58,10 @@ const rejectInterceptor = (error: AxiosError<ApiResponse<ErrorData>>) => {
         return Promise.reject(new FetchError())
     }
 
+    if (error.response?.status === 500) {
+        return Promise.reject(new FetchError())
+    }
+
     return Promise.reject(error)
 }
 
@@ -67,20 +71,27 @@ const client = axios.create({
 
 client.interceptors.response.use(responseInterceptor, rejectInterceptor)
 
-export const postAuthorized = <T, U>(url: string, req: T, token?: string, headers?: AxiosRequestHeaders) =>
-    client.post<U>(url, req, {
+export const postAuthorized = <T, U>(url: string, req: T, token?: string, headers?: AxiosRequestHeaders) => {
+    if (!token) {
+        return Promise.reject(new AuthorizationError())
+    }
+    return client.post<U>(url, req, {
         headers: {
             Authorization: `Bearer ${token}`,
             ...headers,
         },
     })
+}
 
-export const deleteAuthorized = <T, U>(url: string, data: T, token?: string) =>
-    axios.delete<ApiResponse<U>>(url, {
+export const deleteAuthorized = <T, U>(url: string, data: T, token?: string) => {
+    if (!token) {
+        return Promise.reject(new AuthorizationError())
+    }
+    return axios.delete<ApiResponse<U>>(url, {
         data,
         headers: {
             Authorization: `Bearer ${token}`,
         },
     })
-
+}
 export default client
